+++ CHANGELOG.md
@@ -259,3 +259,4 @@
 - GP2-706 - 
 - ablity-to-disable-export-sections
 - GP2-655 - local development support/skeleton for Wagtail Transfer work
+- GP2-667 - appropriately scoped configuration for Wagtail Transfer when deployed